+++ rust/processor/src/db/postgres/migrations/2024-10-02-142213_update_user_pools_with_daily_vol/down.sql
@@ -1,7 +1,7 @@
 -- This file should undo anything in `up.sql`
 DROP FUNCTION user_pools(provider text);
 CREATE FUNCTION user_pools(provider text) RETURNS TABLE(
-  transaction_version NUMERIC,
+  transaction_version BIGINT,
   sender VARCHAR(66),
   entry_function VARCHAR(200),
   transaction_timestamp TIMESTAMP,
